% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UStoEU.R
\name{print.UStoEU}
\alias{print.UStoEU}
\title{Print UStoEU}
\usage{
\method{print}{UStoEU}(x, from, to, return = "value", ...)
}
\arguments{
\item{x}{numeric value}

\item{from}{possible values to choose from: 'ft, 'in', 'm', 'cm', 'lbs'}

\item{to}{possible values to choose from: 'ft, 'in', 'm', 'cm', 'lbs'}

\item{return}{possible values to choose from: 'value', 'ft&in'}

\item{...}{other parameters}
}
\description{
Print the results
}
\author{
Olga Chwedoruk, Tomasz Bugajski
}
