% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UStoEU.R
\name{summary.UStoEU}
\alias{summary.UStoEU}
\title{Summarise UStoEU}
\usage{
\method{summary}{UStoEU}(object, ...)
}
\arguments{
\item{object}{numeric value}

\item{...}{Other parameters}
}
\description{
Summarise the results
}
\author{
Olga Chwedoruk, Tomasz Bugajski
}
