% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UStoEU.R
\name{UStoEU}
\alias{UStoEU}
\title{Converts US units to European and vice versa}
\usage{
UStoEU(x, from, to, return = "value", ...)
}
\arguments{
\item{x}{numeric value}

\item{from}{possible values to choose from: 'ft, 'in', 'm', 'cm', 'lbs'}

\item{to}{possible values to choose from: 'ft, 'in', 'm', 'cm', 'lbs'}

\item{return}{possible values to choose from: 'value', 'ft&in'}

\item{...}{other parameters}
}
\value{
Function returns a measure in the selected unit
}
\description{
The package converts American units such as 'ft, 'in', 'lbs' to European units: 'm', 'cm', 'kg' and vice versa. 'ft' can also be written as 'foot' and 'feet'. 'in' can also be written as 'inch' and 'inches'. 'm' can also be written as 'meter' and 'metre'. 'cm' can also be written as 'centimeter' and 'centimetre'.
}
\author{
Olga Chwedoruk, Tomasz Bugajski
}
